pyproject.toml
[tool.poetry]
name = "stellar-poe"
version = "0.1.0"
description = ""
authors = ["Erick Navarro <erick@navarro.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
gunicorn = "^19.9"
psycopg2-binary = "^2.7"
django-crispy-forms = "^1.7"
django = "^2.1"
django-model-utils = "^3.1"
raven = "^6.10"
dj-database-url = "^0.5.0"
stellar-base = "^1.0"
mnemonic = "^0.18.0"
django-minio-storage = "^0.2.2"

[tool.poetry.dev-dependencies]
pep8-naming = "^0.7.0"
pytest-coverage = "^0.0.0"
mixer = "^6.1"
black = "^18.3-alpha.4"
pytest-django = "^3.4"
flake8 = "^3.6"

[tool.black]
py36 = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
